automation/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-telegram-bot>=20.7
